Move chat non-stream collection to processing.rs
# Conflicts:
#	sgl-router/src/routers/grpc/pipeline.rs
